migrating server logs to tracer
