+++ app/models/shop.ts
@@ -1,206 +1,17 @@
 import PokemonFactory from "./pokemon-factory"
 import { Pkm, PkmFamily } from "../types/enum/Pokemon"
 import Player from "./colyseus-models/player"
-import { Probability, DITTO_RATE } from "../types/Config"
+import { Probability, DITTO_RATE, PoolSize, CommonShop, EpicShop, LegendaryShop, RareShop, UncommonShop } from "../types/Config"
 import { Rarity } from "../types/enum/Game"
 import { pickRandomIn, shuffleArray } from "../utils/random"
 import { clamp } from "../utils/number"
 import { removeInArray } from "../utils/array"
 import { Pokemon } from "./colyseus-models/pokemon"
 
-export const PoolSize: { [key in Rarity]: [number, number, number] } = {
-  [Rarity.COMMON]: [1, 14, 29],
-  [Rarity.UNCOMMON]: [1, 11, 22],
-  [Rarity.RARE]: [1, 9, 18],
-  [Rarity.EPIC]: [1, 7, 14],
-  [Rarity.LEGENDARY]: [1, 5, 10],
-  [Rarity.MYTHICAL]: [1, 1, 1],
-  [Rarity.NEUTRAL]: [0, 0, 0],
-  [Rarity.SUMMON]: [0, 0, 0],
-  [Rarity.HATCH]: [0, 0, 0]
-}
 
 export function getPoolSize(rarity: Rarity, maxStars: number): number {
   return PoolSize[rarity][clamp(maxStars, 1, 3) - 1]
 }
-
-export const CommonShop = new Array<Pkm>(
-  Pkm.POLIWAG,
-  Pkm.CHARMANDER,
-  Pkm.GEODUDE,
-  Pkm.AZURILL,
-  Pkm.ZUBAT,
-  Pkm.MAREEP,
-  Pkm.CLEFFA,
-  Pkm.CATERPIE,
-  Pkm.WEEDLE,
-  Pkm.PIDGEY,
-  Pkm.HOPPIP,
-  Pkm.SEEDOT,
-  Pkm.STARLY,
-  Pkm.SWINUB,
-  Pkm.FENNEKIN,
-  Pkm.PICHU,
-  Pkm.SQUIRTLE,
-  Pkm.ARON,
-  Pkm.MUDKIP,
-  Pkm.CHIMCHAR
-)
-
-export const UncommonShop = new Array<Pkm>(
-  Pkm.TOTODILE,
-  Pkm.IGGLYBUFF,
-  Pkm.CHIKORITA,
-  Pkm.CYNDAQUIL,
-  Pkm.TREECKO,
-  Pkm.TORCHIC,
-  Pkm.PIPLUP,
-  Pkm.MACHOP,
-  Pkm.HORSEA,
-  Pkm.SPHEAL,
-  Pkm.MAGNEMITE,
-  Pkm.DUSKULL,
-  Pkm.EEVEE,
-  Pkm.FLABEBE,
-  Pkm.BELLSPROUT,
-  Pkm.SLOWPOKE,
-  Pkm.HATENNA,
-  Pkm.NIDORANF,
-  Pkm.NIDORANM
-)
-
-export const RareShop = new Array<Pkm>(
-  Pkm.BULBASAUR,
-  Pkm.TURTWIG,
-  Pkm.TRAPINCH,
-  Pkm.RHYHORN,
-  Pkm.TOGEPI,
-  Pkm.LOTAD,
-  Pkm.SHINX,
-  Pkm.DRATINI,
-  Pkm.MAGBY,
-  Pkm.WHISMUR,
-  Pkm.VANILLITE,
-  Pkm.BAGON,
-  Pkm.HONEDGE,
-  Pkm.ABRA
-)
-
-export const EpicShop = new Array<Pkm>(
-  Pkm.LARVITAR,
-  Pkm.SLAKOTH,
-  Pkm.RALTS,
-  Pkm.BELDUM,
-  Pkm.GIBLE,
-  Pkm.ELEKID,
-  Pkm.SNORUNT,
-  Pkm.BUDEW,
-  Pkm.PORYGON,
-  Pkm.CUBONE,
-  Pkm.HOUNDOUR,
-  Pkm.GOOMY,
-  Pkm.BOUNSWEET,
-  Pkm.OSHAWOTT,
-  Pkm.JANGMO_O
-)
-
-export const LegendaryShop = new Array<Pkm>(
-  Pkm.GASTLY,
-  Pkm.ONIX,
-  Pkm.SCYTHER,
-  Pkm.RIOLU,
-  Pkm.MEDITITE,
-  Pkm.NUMEL,
-  Pkm.SNOVER,
-  Pkm.SWABLU,
-  Pkm.BUNEARY,
-  Pkm.ELECTRIKE,
-  Pkm.SHUPPET,
-  Pkm.NINCADA,
-  Pkm.HAPPINY,
-  Pkm.SOLOSIS
-)
-
-export const Mythical1Shop = new Array<Pkm>(
-  Pkm.AERODACTYL,
-  Pkm.BLACEPHALON,
-  Pkm.REGIDRAGO,
-  Pkm.REGIELEKI,
-  Pkm.CASTFORM,
-  Pkm.REGICE,
-  Pkm.REGISTEEL,
-  Pkm.REGIROCK,
-  Pkm.UXIE,
-  Pkm.MESPRIT,
-  Pkm.AZELF,
-  Pkm.LATIAS,
-  Pkm.LATIOS,
-  Pkm.ZAPDOS,
-  Pkm.MOLTRES,
-  Pkm.ARTICUNO,
-  Pkm.LAPRAS,
-  Pkm.ABSOL,
-  Pkm.SPIRITOMB,
-  Pkm.ROTOM,
-  Pkm.MANAPHY,
-  Pkm.COBALION,
-  Pkm.KELDEO,
-  Pkm.VOLCARONA,
-  Pkm.TAPU_KOKO,
-  Pkm.TAPU_LELE,
-  Pkm.SEVIPER,
-  Pkm.KECLEON,
-  Pkm.MAWILE,
-  Pkm.TAUROS,
-  Pkm.TORNADUS,
-  Pkm.RELICANTH,
-  Pkm.MEW,
-  Pkm.CHATOT,
-  Pkm.FARFETCH_D,
-  Pkm.TAPU_BULU,
-  Pkm.TAPU_FINI,
-  Pkm.MIMIKYU,
-  Pkm.TYROGUE,
-  Pkm.VIRIZION,
-  Pkm.ZERAORA
-)
-
-export const Mythical2Shop = new Array<Pkm>(
-  Pkm.RESHIRAM,
-  Pkm.ZEKROM,
-  Pkm.STAKATAKA,
-  Pkm.GENESECT,
-  Pkm.GUZZLORD,
-  Pkm.ETERNATUS,
-  Pkm.MELOETTA,
-  Pkm.MEWTWO,
-  Pkm.ENTEI,
-  Pkm.SUICUNE,
-  Pkm.RAIKOU,
-  Pkm.REGIGIGAS,
-  Pkm.CELEBI,
-  Pkm.VICTINI,
-  Pkm.JIRACHI,
-  Pkm.ARCEUS,
-  Pkm.DEOXYS,
-  Pkm.SHAYMIN,
-  Pkm.GIRATINA,
-  Pkm.DARKRAI,
-  Pkm.CRESSELIA,
-  Pkm.HEATRAN,
-  Pkm.LUGIA,
-  Pkm.HO_OH,
-  Pkm.PALKIA,
-  Pkm.DIALGA,
-  Pkm.RAYQUAZA,
-  Pkm.KYOGRE,
-  Pkm.GROUDON,
-  Pkm.VOLCANION,
-  Pkm.ORIGIN_GIRATINA,
-  Pkm.MARSHADOW,
-  Pkm.XERNEAS,
-  Pkm.YVELTAL
-)
 
 export default class Shop {
   commonPool: Map<Pkm, number> = new Map<Pkm, number>()
@@ -400,7 +211,7 @@ export default class Shop {
             pokemon = this.getRandomPokemonFromPool(this.legendaryPool, finals)
             break
           default:
-            console.log(
+            console.error(
               `error in shop while picking seed = ${seed}, threshold = ${threshold}, index = ${i}`
             )
             break
