+++ app/utils/logger.ts
@@ -1,0 +1,5 @@
+export const logger = require("tracer").dailyfile({
+    root: 'logs',
+    maxLogFiles: 10,
+    allLogsFileName: 'pac'
+})
